% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_empty}
\alias{clean_empty}
\title{Cleaning empty results}
\usage{
clean_empty(inventory_data, cols)
}
\arguments{
\item{inventory_data}{Data from an inventory as obtained from
\code{\link{ifn_to_tibble}}, \code{\link{fia_to_tibble}} or
\code{\link{ffi_to_tibble}}.}

\item{cols}{vector with column names to clean from empty
results. Can be one or more of \code{"tree"},
\code{"understory"} and \code{"regen"}. If more than one,
only plots with data in all columns selected will be
retained.}
}
\value{
A tibble the same as \code{inventory_data} with the
empty data removed for the columns selected.
}
\description{
Cleaning inventory results to filter out empty data
}
\details{
This functions remove plot rows with empty data in the desired
nested columns
}
