% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_filter_list}
\alias{create_filter_list}
\title{Create a compatible filter_list object}
\usage{
create_filter_list(plots_info)
}
\arguments{
\item{plots_info}{Object resulted from \code{link{show_plots_from}}, or a compatible one
(\emph{i.e.} the same object after some plot filtering).}
}
\value{
A list object compatible with the \code{filter_list} argument of
\code{\link{ffi_to_tibble}}, \code{\link{fia_to_tibble}} or \code{\link{ifn_to_tibble}}
}
\description{
Create a compatible filter_list object from the result of \code{link{show_plots_from}}
}
\details{
This function takes the result of \code{link{show_plots_from}}, or a compatible object and
creates a \code{filter_list} object ready to be use with \code{\link{ffi_to_tibble}},
\code{\link{fia_to_tibble}} or \code{\link{ifn_to_tibble}}. Internal heuristics determine the
inventory from the data supplied.
}
\examples{

library(esus)

# FIA
show_plots_from("FIA", folder = ".", states = "OR") |>
  create_filter_list()
# FFI
show_plots_from("FFI", folder = ".", departments = "21") |>
  create_filter_list()
# IFN
show_plots_from("IFN", folder = ".", provinces = "24", version = "ifn4") |>
  create_filter_list()

}
