% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi_utils.R
\name{create_filter_list_ffi}
\alias{create_filter_list_ffi}
\title{Create the \code{filter_list} for FFI inventory}
\usage{
create_filter_list_ffi(plots_info)
}
\description{
Create the \code{filter_list} for FFI inventory
}
