% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_utils.R
\name{create_filter_list_fia}
\alias{create_filter_list_fia}
\title{Create the \code{filter_list} for FIA inventory}
\usage{
create_filter_list_fia(plots_info)
}
\description{
Create the \code{filter_list} for FIA inventory
}
