% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ffi_output_example}
\alias{ffi_output_example}
\title{FFI example output}
\format{
\subsection{\code{ffi_output_example}}{

A nested dataframe with the FFI plots for the selected years in the \emph{Loire}
department
}
}
\source{
FFI raw data obtained from \url{https://inventaire-forestier.ign.fr/}
}
\usage{
ffi_output_example
}
\description{
A subset of plots from the \emph{Loire} department for 2015 to 2018.
}
\keyword{datasets}
