% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fia_output_example}
\alias{fia_output_example}
\title{FIA example output}
\format{
\subsection{\code{fia_output_example}}{

A nested dataframe with the FIA plots for the selected years in Alaska
}
}
\source{
FIA raw data obtained from \url{https://www.fs.usda.gov/research/programs/nfi#data-and-tools}
}
\usage{
fia_output_example
}
\description{
A subset of plots from Alaska (AK) for 2015 to 2018.
}
\keyword{datasets}
