% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifn_output_example}
\alias{ifn_output_example}
\title{IFN example output}
\format{
\subsection{\code{ifn_output_example}}{

A nested dataframe with the IFN plots for the selected versions in \emph{Barcelona} province
}
}
\source{
IFN raw data obtained from
\url{https://www.miteco.gob.es/es/biodiversidad/temas/inventarios-nacionales/inventario-forestal-nacional/cuarto_inventario.html}
}
\usage{
ifn_output_example
}
\description{
A subset of plots from \emph{Barcelona} province for the IFN2, IFN3 and IFN4 versions.
}
\keyword{datasets}
