% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inventory_as_sf}
\alias{inventory_as_sf}
\title{Convert inventory data to sf based on coords and crs present}
\usage{
inventory_as_sf(inventory_data)
}
\arguments{
\item{inventory_data}{Data from an inventory as obtained from
\code{\link{ifn_to_tibble}}, \code{\link{fia_to_tibble}} or
\code{\link{ffi_to_tibble}}.}
}
\value{
An sf object with the same data as \code{inventory_data}
and a new column with the original crs for traceability.
}
\description{
Use coords vars and crs to convert to sf
}
