% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_plots_from}
\alias{show_plots_from}
\title{Show plots for any inventory}
\usage{
show_plots_from(inventory = c("FIA", "FFI", "IFN"), folder = ".", ...)
}
\arguments{
\item{inventory}{Character indicating the inventory. Allowed values are \code{"FIA"} for the
USA forest inventory, \code{"FFI"} for the French \emph{Inventaire Forestier} and
\code{"IFN"}, for the Spanish \emph{Inventario Forestal Nacional}.}

\item{folder}{Character, path to the folder containing the \code{inventory} files.}

\item{...}{Other arguments, depending on the \code{inventory}, see inventory sections.}
}
\value{
A \code{\link[sf]{sf}} spatial object in which each row is a plot,. The metadata provided
varies depending on the inventory, but usually includes the state (FIA) / department (FFI)/
provincia (IFN) and year/date/IFN version
}
\description{
Show plots with minimal metadata from any inventory
}
\details{
This function show the plots available in any inventory for the given administrative units.
Take into account that this can potentially show all plots in any inventory, so the object
returned can be memory heavy.
}
\section{FIA}{

FIA needs an extra argument, \code{states}, a character vector with the two-letter
code for the desired states.
}

\section{FFI}{

FFI needs an extra argument, \code{departments}, a character vector with the desired
department codes.
}

\section{IFN}{

IFN needs two extra arguments, \code{provinces}, a character vector with the numeric codes for
the provinces and \code{version}, a character with the IFN version to look at (\code{"ifn2"},
\code{"ifn3"} or \code{"ifn4"}).
}

\examples{
library(esus)

# FIA
show_plots_from("FIA", folder = ".", states = "OR")
# FFI
show_plots_from("FFI", folder = ".", departments = "21")
# IFN
show_plots_from("IFN", folder = ".", provinces = "24", version = "ifn4")

}
