% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R, R/fia.R
\name{tables_processing}
\alias{tables_processing}
\alias{ffi_plot_table_process}
\alias{ffi_tree_table_process}
\alias{ffi_shrub_table_process}
\alias{fia_plot_table_process}
\alias{fia_tree_table_process}
\alias{fia_p3_understory_table_process}
\alias{fia_p2_understory_table_process}
\alias{fia_seedling_table_process}
\alias{fia_subplot_table_process}
\alias{fia_soils_lab_table_process}
\alias{fia_soils_loc_table_process}
\title{Data tables process}
\usage{
ffi_plot_table_process(plot_data, soil_data, plot, year, def_metadonnes)

ffi_tree_table_process(tree_data, plot, year, espar_cdref13, espar_ref)

ffi_shrub_table_process(
  shrub_data,
  plot,
  year,
  cd_ref,
  growth_form_lignified_france
)

fia_plot_table_process(plot_data, survey_data, cond_data, plot, county, year)

fia_tree_table_process(tree_data, plot, county, year, ref_species)

fia_p3_understory_table_process(
  understory_data,
  plot,
  county,
  year,
  growth_habit,
  ref_plant_dictionary
)

fia_p2_understory_table_process(
  understory_p2,
  plot,
  county,
  year,
  growth_habit,
  ref_plant_dictionary
)

fia_seedling_table_process(seedling_data, plot, county, year, ref_species)

fia_subplot_table_process(subplot_data, plot, county, year)

fia_soils_lab_table_process(soils_lab, plot, county, state, year)

fia_soils_loc_table_process(soils_loc, veg_subplot, plot, county, state, year)
}
\arguments{
\item{plot_data, tree_data, shrub_data, soil_data}{Paths to the files with the corresponding data}

\item{plot}{Numeric, plot code}

\item{year}{Numeric, year to extract}

\item{espar_cdref13, cdref13, ESPAR_REF, def_metadonnes, growth_form_lignified_france}{tables. These tables
are automatically read in \code{\link{fia_tables_process}} based on the folder provided.}

\item{county}{COUNTYCD code}

\item{ref_species, ref_plant_dictionary}{ref species and ref plant dictionary tables. These tables
are automatically read in \code{\link{fia_tables_process}} based on the folder provided.}

\item{growth_habit}{Character, growth habit value to filter data (to distinguish between herbs
and shrubs)}

\item{dep}{department code}

\item{plot_data, survey_data, cond_data, tree_data, understory_data, understory_p2, seedling_data, subplot_data, soils_lab, soils_loc}{Paths to the files with the corresponding data}
}
\value{
A tibble with one or more rows (depending on the data retrieved) for each plot for that
year.

A tibble with one or more rows (depending on the data retrieved) for each plot for that
year.
}
\description{
Process to gather needed data from FFI csv tables

Process to gather needed data from FIA csv tables
}
\details{
These functions retrieve the data for a plot in one year.

These functions retrieve the data for a plot in one year.
}
\section{Functions}{
\itemize{
\item \code{ffi_plot_table_process()}: Process to gather needed data from plot, survey and cond tables

\item \code{ffi_tree_table_process()}: Process to gather needed data from tree table

\item \code{ffi_shrub_table_process()}: Process to gather needed data from xxxxx

\item \code{fia_plot_table_process()}: Process to gather needed data from plot, survey and cond tables

\item \code{fia_tree_table_process()}: Process to gather needed data from tree table

\item \code{fia_p3_understory_table_process()}: Process to gather needed data from veg subplot spp table

\item \code{fia_p2_understory_table_process()}: Process to gather needed data from p2 veg subplot spp table

\item \code{fia_seedling_table_process()}: Process to gather needed data from seedling table

\item \code{fia_subplot_table_process()}: Process to gather needed data from subplot table

\item \code{fia_soils_lab_table_process()}: Process to gather needed data from soils lab table

\item \code{fia_soils_loc_table_process()}: Process to gather needed data from soils sample loc table

}}
