% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_empty}
\alias{clean_empty}
\title{Cleaning empty results}
\usage{
clean_empty(inventory_data, cols)
}
\arguments{
\item{inventory_data}{Data from an inventory as obtained from \code{\link{ifn_to_tibble}},
\code{\link{fia_to_tibble}} or \code{\link{ffi_to_tibble}}.}

\item{cols}{vector with column names to clean from empty results. Can be one or more of
\code{"tree"}, \code{"understory"} and \code{"regen"}. If more than one, only plots with
data in all columns selected will be retained.}
}
\value{
A tibble the same as \code{inventory_data} with the empty data removed for the columns
selected.
}
\description{
Cleaning inventory results to filter out empty data
}
\details{
This functions remove plot rows with empty data in the desired nested columns.
}
\examples{
\donttest{
\dontrun{
library(esus)

# FFI
ffi_to_tibble(
  departments = c("01"), years = c(2019),
  filter_list = list("01" = c(1404119)),
  folder = "path/to/ffi/data"
) |>
  clean_empty(c("tree", "regen", "understory"))

# FIA
fia_to_tibble(
  years = 2019, states = c("OR"),
  filter_list = list("OR" = list("59" = c(76413))),
  folder = "path/to/fia/data"
) |>
  clean_empty(c("tree", "regen", "understory"))

# IFN
ifn_to_tibble(
  provinces = c("24"), versions = c("ifn3"),
  filter_list = list("24" = c("24_0270_xx_A4_xx")),
  folder = "path/to/ifn/data"
) |>
  clean_empty(c("tree", "regen", "understory"))
}
}

}
