% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_inventory}
\alias{download_inventory}
\title{Download the especified inventory}
\usage{
download_inventory(
  inventory = c("fia", "ffi", "ifn"),
  destination = ".",
  .verbose = TRUE
)
}
\arguments{
\item{inventory}{Character with the inventory abbreviation}

\item{destination}{Path to the inventory destination folder. This folder must exists.}

\item{.verbose}{Logical indicating if progress messages should be shown.}
}
\value{
Invisible TRUE if the download and unzip was succesful, an error otherwise.
}
\description{
Download and unzip the inventory data files at the desired destination
}
\details{
This function tries to download the available files for the especified inventory
from the official repositories of the inventory. It can fail as can be connection
problems or files missing temporary (This usually happens for the version 4 of the IFN)
}
\examples{
\donttest{
\dontrun{
library(forestables)
download_inventory(ffi, destination = tempdir())
}
}
}
