% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestables-package.R
\docType{package}
\name{forestables-package}
\alias{forestables}
\alias{forestables-package}
\title{forestables: Making Forest Inventory Data Tidy}
\description{
The `forestables` package converts raw forest inventory files from the Spanish Forest Inventory (IFN), the French Forest Inventory (FFI) and the USA Forest Inventory (FIA) into tidy tables to be used in analyses and modelling.
}
\author{
\strong{Maintainer}: Víctor Granda \email{victorgrandagarcia@gmail.com} (\href{https://orcid.org/0000-0002-0469-1991}{ORCID})

Authors:
\itemize{
  \item Adriana Tovar [contributor]
}

Other contributors:
\itemize{
  \item Miquel de Cáceres (\href{https://orcid.org/0000-0001-7132-2080}{ORCID}) [contributor]
}

}
\keyword{internal}
