% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R, R/fia.R
\name{tables_processing}
\alias{tables_processing}
\alias{ffi_plot_table_process}
\alias{ffi_tree_table_process}
\alias{ffi_shrub_table_process}
\alias{ffi_regen_table_process}
\alias{fia_plot_table_process}
\alias{fia_tree_table_process}
\alias{fia_understory_table_process}
\alias{fia_p3_understory_table_process}
\alias{fia_p2_understory_table_process}
\alias{fia_seedling_table_process}
\alias{fia_subplot_table_process}
\title{Data tables process}
\usage{
ffi_plot_table_process(plot_data, soils_data, plot, year, metadonnees)

ffi_tree_table_process(tree_data, plot, year, espar_cdref, idp_dep_ref)

ffi_shrub_table_process(
  shrub_data,
  plot,
  year,
  cd_ref,
  growth_form_lignified_france,
  idp_dep_ref
)

ffi_regen_table_process(regen_data, plot, year, espar_cdref, idp_dep_ref)

fia_plot_table_process(plot_data, survey_data, cond_data, plot, county, year)

fia_tree_table_process(tree_data, plot, county, year, ref_species)

fia_understory_table_process(
  understory_data,
  understory_p2,
  plot,
  county,
  year,
  growth_habit_p3,
  growth_habit_p2,
  ref_plant_dictionary
)

fia_p3_understory_table_process(
  understory_data,
  plot,
  county,
  year,
  growth_habit,
  ref_plant_dictionary
)

fia_p2_understory_table_process(
  understory_p2,
  plot,
  county,
  year,
  growth_habit,
  ref_plant_dictionary
)

fia_seedling_table_process(seedling_data, plot, county, year, ref_species)

fia_subplot_table_process(subplot_data, plot, county, year)
}
\arguments{
\item{plot_data, tree_data, shrub_data, soils_data}{Paths to the files with the corresponding data}

\item{plot}{Numeric, plot code}

\item{year}{Numeric, year to extract}

\item{espar_cdref, metadonnees, growth_form_lignified_france}{tables. These tables
are automatically read in \code{\link{ffi_tables_process}} based on the folder provided.}

\item{county}{COUNTYCD code}

\item{ref_species, ref_plant_dictionary}{ref species and ref plant dictionary tables. These tables
are automatically read in \code{\link{fia_tables_process}} based on the folder provided.}

\item{growth_habit}{Character, growth habit value to filter data (to distinguish between herbs
and shrubs)}

\item{dep}{department code}

\item{plot_data, survey_data, cond_data, tree_data, understory_data, understory_p2, seedling_data, subplot_data, soils_lab, soils_loc}{Paths to the files with the corresponding data}
}
\value{
A tibble with one or more rows (depending on the data retrieved) for each plot for that
year.

A tibble with one or more rows (depending on the data retrieved) for each plot for that
year.
}
\description{
Process to gather needed data from FFI csv tables

Process to gather needed data from FIA csv tables
}
\details{
These functions retrieve the data for a plot in one year.

These functions retrieve the data for a plot in one year.
}
\section{Functions}{
\itemize{
\item \code{ffi_plot_table_process()}: Process to gather needed data from plot, survey and cond tables

\item \code{ffi_tree_table_process()}: Process to gather needed data from tree table

\item \code{ffi_shrub_table_process()}: Process to gather needed data from shrub table

\item \code{ffi_regen_table_process()}: Process to gather needed data from soil table

\item \code{fia_plot_table_process()}: Process to gather needed data from plot, survey and cond tables

\item \code{fia_tree_table_process()}: Process to gather needed data from tree table

\item \code{fia_understory_table_process()}: Process to guess which understory data is available and launch the
corresponding function.

\item \code{fia_p3_understory_table_process()}: Process to gather needed data from veg subplot spp table

\item \code{fia_p2_understory_table_process()}: Process to gather needed data from p2 veg subplot spp table

\item \code{fia_seedling_table_process()}: Process to gather needed data from seedling table

\item \code{fia_subplot_table_process()}: Process to gather needed data from subplot table

}}
